+++ agent/internal/ebpf/probes/c/openat_monitor.c
@@ -19,11 +19,6 @@
 #include "common.h"
 #include <bpf/bpf_helpers.h>
 
-struct {
-    __uint(type, BPF_MAP_TYPE_RINGBUF);
-    __uint(max_entries, 1 << 18);
-} events SEC(".maps");
-
 SEC("tracepoint/syscalls/sys_enter_openat")
 int handle_openat(struct trace_event_raw_sys_enter *ctx) {
     struct bpf_event *e = bpf_ringbuf_reserve(&events, sizeof(*e), 0);
@@ -45,5 +40,3 @@ int handle_openat(struct trace_event_raw_sys_enter *ctx) {
     bpf_ringbuf_submit(e, 0);
     return 0;
 }
-
-char _license[] SEC("license") = "GPL";