+++ agent/internal/ebpf/probes/c/setuid_monitor.c
@@ -16,11 +16,6 @@
 #include "common.h"
 #include <bpf/bpf_helpers.h>
 
-struct {
-    __uint(type, BPF_MAP_TYPE_RINGBUF);
-    __uint(max_entries, 1 << 18);
-} events SEC(".maps");
-
 SEC("tracepoint/syscalls/sys_enter_setuid")
 int handle_setuid(struct trace_event_raw_sys_enter *ctx) {
     struct bpf_event *e = bpf_ringbuf_reserve(&events, sizeof(*e), 0);
@@ -39,5 +34,3 @@ int handle_setuid(struct trace_event_raw_sys_enter *ctx) {
     bpf_ringbuf_submit(e, 0);
     return 0;
 }
-
-char _license[] SEC("license") = "GPL";